Add `IdentityWrapper` component.
Feedback would be lovely.

The basis of this component is that it will wrap other components, and
ensure that they have a `getValue()` method associated with
them. Currently, it acts as a pass-through for components that already
implement their own `getValue()` functions, and handles basic DOM form
components (input, select, textarea).

Usage of this compnent can be seen in the test cases.

- - -

Implementation:

So, this is a bit of a woozy and it's due to the way input type radio is
implemented (and possibly others?). Instead of accepting only one
child for IdentityWrapper, we now have to consider multiple
children. e.g.

    <IdentityWrapper>
        <input type="radio" name="foo" value="cat" />
        <input type="radio" name="foo" value="cat" />
    </IdentityWrapper>

Because of this, the implementation maps over `this.refs`, and pulls out
the values of each child. If it is an invalid node passed in, it returns
undefined as a value for that child. In the case of radio inputs, it
returns undefined for unchecked radio buttons and the value of the radio
button if it is checked.

After we have an array of values, filter out the undefined values. There
should only be *zero or one* values in the array after filtering out
undefined values. If zero, return undefined. If one, return that element
from the values array (`values[0]`).

The implementation of IdentityWrapper can only return one value from
`getValue()`. A warning is issued if something like this is passed in:

    <IdentityWrapper>
        <input type="text" name="foo" />
        <input type="text" name="bar" />
    </IdentityWrapper>

There are other warnings in place to help indicate to users why their
input is returning undefined. Should we throw an Error vs. a warning in
these cases?

- - -

This is a rough implementation but I _think_ it covers the majority of
use cases. Have I missed any?